+++ R/accounts.R
@@ -278,9 +278,9 @@ getAuthedUser <- function(
   client$currentUser()
 }
 
-#' Register account on shinyapps.io or posit.cloud
+#' Register account on shinyapps.io
 #'
-#' Configure a ShinyApps or Posit Cloud account for publishing from this system.
+#' Configure a ShinyApps account for publishing from this system.
 #'
 #' @param name Name of account to save or remove
 #' @param token User token for the account