+++ R/client.R
@@ -5,8 +5,6 @@ clientForAccount <- function(account) {
 
   if (isShinyappsServer(account$server)) {
     shinyAppsClient(serverUrl, account)
-  } else if (isPositCloudServer(account$server)) {
-    cloudClient(serverUrl, account)
   } else if (isSPCSServer(account$server)) {
     account$snowflakeToken <- getSnowflakeAuthToken(
       serverInfo$url,
@@ -123,7 +121,8 @@ isContentType <- function(x, contentType) {
   grepl(contentType, x, fixed = TRUE)
 }
 
-uploadCloudBundle <- function(
+
+uploadShinyappsBundle <- function(
   client,
   application_id,
   bundlePath,
@@ -152,6 +151,7 @@ uploadCloudBundle <- function(
   # Step 4. Retrieve updated bundle post status change
   client$getBundle(bundle$id)
 }
+
 
 uploadBundle <- function(bundle, bundleSize, bundlePath) {
   presigned_service <- parseHttpUrl(bundle$presigned_url)