+++ NEWS.md
@@ -15,6 +15,10 @@
 
 * Improve documentation and advice for `deployApp(envVars...)`.
 
+* Removed support for publishing to Posit Cloud. (#1215)
+
+* Removed the Cloud-exclusive `space` argument from `deployApp()`. (#1215)
+
 # rsconnect 1.5.1
 
 * Address user registration for Posit Connect deployments hosted in Snowpark